move source types to types dir
