+++ src/store/source/showProperties.ts
@@ -3,8 +3,8 @@ import { createAsyncThunk } from "@reduxjs/toolkit"
 import { Id, SourceType } from "@/types"
 import { emit } from "@tauri-apps/api/event"
 import { RootState } from ".."
-import { Source } from "."
 import { waitEvent } from "@/lib/tauri"
+import type { Source } from "@/types/source"
 import { SourceReader } from "@/lib/source-reader"
 import logger from "@/logger"
